backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0

# Anthropic SDK for agents
anthropic>=0.40.0

# HTTP client
httpx==0.26.0

# Data processing
pandas==2.1.4
python-dateutil==2.8.2

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Server-Sent Events
sse-starlette==1.8.2
